# Conflicts:
#	customerapp/proguard-rules.txt
#	features/discovery/home/src/main/java/com/tokopedia/home/beranda/presentation/view/adapter/viewholder/static_channel/OvoViewHolder.kt
#	features/discovery/search/src/main/java/com/tokopedia/search/result/presentation/view/adapter/viewholder/product/InspirationCarouselViewHolder.kt
#	features/entertainment/digital_deals/src/main/java/com/tokopedia/digital_deals/view/activity/AllBrandsActivity.java
#	features/entertainment/digital_deals/src/main/java/com/tokopedia/digital_deals/view/activity/DealsSearchActivity.java
#	features/entertainment/digital_deals/src/main/java/com/tokopedia/digital_deals/view/adapter/TrendingDealsAdapter.java
#	features/entertainment/digital_deals/src/main/java/com/tokopedia/digital_deals/view/fragment/CategoryDetailHomeFragment.java
#	features/entertainment/digital_deals/src/main/java/com/tokopedia/digital_deals/view/fragment/DealsHomeFragment.java
#	features/entertainment/digital_deals/src/main/res/values/dimens.xml
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/common/util/AddEditProductNotificationManager.kt
#	features/promo/promo_checkout/src/main/java/com/tokopedia/promocheckout/list/view/adapter/PromoCheckOutExchangeCouponAdapter.kt
#	features/promo/tokopoints/src/main/java/com/tokopedia/tokopoints/view/customview/ServerErrorView.java
#	features/promo/tokopoints/src/main/java/com/tokopedia/tokopoints/view/tokopointhome/catalog/SectionHorizontalCatalogVH.kt
#	features/recharge/digitalproduct/src/main/java/com/tokopedia/digital/newcart/presentation/fragment/DigitalCartDealsFragment.java
#	features/recharge/digitalproduct/src/main/java/com/tokopedia/digital/newcart/presentation/fragment/DigitalDealCheckoutFragment.java
#	features/recharge/digitalproduct/src/main/java/com/tokopedia/digital/product/view/fragment/DigitalProductFragment.java
#	features/recharge/digitalproduct/src/main/java/com/tokopedia/digital/product/view/fragment/DigitalUssdFragment.java
#	features/recharge/digitalproduct/src/main/res/values/dimens.xml
#	features/user/loginregister/src/main/java/com/tokopedia/loginregister/activation/view/fragment/ActivationFragment.java
#	features/user/topchat/src/main/java/com/tokopedia/topchat/chattemplate/view/fragment/EditTemplateChatFragment.java
#	libraries/utils/src/main/java/com/tokopedia/utils/permission/SlicePermission.kt

+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/domain/model/response/atcexternal/AddToCartExternalDataModel.kt
@@ -7,12 +7,12 @@ data class AddToCartExternalModel(
 )
 
 data class AddToCartExternalDataModel(
-        var productId: Int = 0,
+        var productId: Long = 0,
         var productName: String = "",
         var quantity: Int = 0,
         var price: Int = 0,
         var category: String = "",
-        var shopId: Int = 0,
+        var shopId: Long = 0,
         var shopType: String = "",
         var shopName: String = "",
         var picture: String = "",
